requirements.txt
dspy-ai[qdrant]
jupyter
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
transformers==4.39.0